+++ src/bgen_prog.cpp
@@ -40,7 +40,9 @@ int main( int argc, char** argv ) {
 		data.standardise_non_genetic_data();
 
 		data.read_full_bgen();
-		data.calc_dxteex();
+		if(data.n_effects > 1){
+			data.calc_dxteex();
+		}
 		if(p.env_weights_file == "NULL" && p.init_weights_with_snpwise_scan) {
 			data.calc_snpstats();
 		}